Community detection on bipartite networks

Run a community detection algorithm on a bipartite network. Algorithms available are from igraph package (greedy, girvan, walktrap, louvain, infomap, spinglass, leading_eigen, label_prop or optimal) and also from the bipartite package (beckett and dorman). Algorithm can be weighted in case of a weighted bipartite network.

algo_bipartite(dat, algo = "greedy", weight = FALSE, input = "matrix",
site = NULL, sp = NULL, ab = NULL)

Arguments

dat
a contingency matrix with sites as rows and species as columns.

algo
An algorithm used to perfom the bipartite community detection. Available options are greedy, girvan, walktrap, louvain, infomap, infomap, spinglass, leading_eigen, label_prop, optimal, beckett and dorman.

weight
A boolean indicating whether the detection of communities has to be weighted or not.

input
A character string indicating whether the input format is a contingency matrix or a long-format data.frame.

site
a character string corresponding to the column with sites.

sp
a character string corresponding to the column with species.

ab
a character string corresponding to the column with abundances of species at sites.

Details

Note

Examples

# Example dataset from bioRgeo:
data("virtual_sp")
sp_df <- virtual[[1]]

# fastgreedy algorithm from igraph
bip_df <- algo_bipartite(dat = sp_df[which(sp_df$pa > 0), ],
                         algo = "greedy",
                         weight = FALSE, input = "data frame", site = "site",
                         sp = "sp", ab = NULL)

#> Le chargement a nécessité le package : igraph

#> Warning: le package 'igraph' a été compilé avec la version R 3.6.2

#> 
#> Attachement du package : 'igraph'

#> The following objects are masked from 'package:stats':
#> 
#>     decompose, spectrum

#> The following object is masked from 'package:base':
#> 
#>     union