Clustering analysis

Assign the sites of a contingency matrix to clusters.

cluster(dat, method = "ward.D2", n_clust = 4)

Arguments

dat
a contingency matrix or a distance matrix.

method
Character string to chose among the following options: kmeans, meanshift, ward.D, ward.D2, single, complete, average, mcquitty, median, centroid, dbscan, gmm, diana or pam.

n_clust
An integer determining the number of clusters. If set to NULL, an optimal number of clusters will be determined.

optim_method
When the chosen method is supervised and there is no a priori regarding the optimal number of clusters, an optimization algorithm is added. The following gap statistic to determine the optimal number of cluster are: globalmax, firstmax, Tibs2001SEmax, firstSEmax or globalSEmax.

nstart
An integer determining the number of random centroids to start K-means analysis and to find the optimal number of clusters.

B
An integer determining the number of Monte Carlo bootstrap samples. Necessary when an algorithm is run to determine the optimal number of clusters.

K.max
A numeric determining the maximum number of clusters to consider. Necessary when an algorithm is run to determine the optimal number of clusters.

Details

The acronym given as argument to the cluster function correspond to the following methods: kmeans => K-Means

pam => Partitioning Around Medoids

meanshift => Mean-Shift

dbscan => Density-Based Spatial Clustering of Applications with Noise

gmm => Gaussian Mixture Models

ward.D => Ward.D

ward.D2 => Ward.D2

single => single

complete => complete

average => Unweighted pair-group method using arithmetic averages (UPGMA, average linkage)

mcquitty => Weighted pair-group method using arithmetic averages (WPGMA, McQuitty s method)

median => Weighted pair-group method using centroids (WPGMC, median linkage)

centroid => Unweighted pair-group method using centroids (UPGMC, centroid linkage)

diana = > DIvisive ANAlysis Clustering

These algorithms were implemented in other R packages, which bioRgeo::cluster() call.

See hclust from stats, diana from cluster, hclust from fastcluster, Mclust from mclust, dbscan from dbscan and meanShift from meanShiftR for further details about the available methods.

Value

A data frame with two columns: the sites (or rows of the contingency table) and the clusters assigned to each site for the chosen method.

Examples

ex <- data.frame(sites = c(rep("A", 2), rep("B", 3), rep("C", 2)),
species = c("a", "b", "a", "c", "d", "b", "d"),
count = c(10, 100, 1, 20, 50, 10, 20))

tmp <- contingency(dat = ex, sp = "species", site = "sites", ab = "count")
cluster(tmp, method = "ward.D2", optim_method = "firstSEmax",
                         nstart = 25, B = 50, K.max = 2)

#> Le chargement a nécessité le package : cluster

#> Warning: One of the chosen method is a supervised algorithm that needs a
#>     number of clusters. Since 'n_clust = NULL', an optimization algorithm is
#>     first executed to determine the optimal numbers of clusters.
#>     This step may take a while.

#> Le chargement a nécessité le package : Rcpp

#> Le chargement a nécessité le package : SMUT

#> Le chargement a nécessité le package : reshape2

#> Warning: le package 'reshape2' a été compilé avec la version R 3.6.3

#> Le chargement a nécessité le package : fastcluster

#> 
#> Attachement du package : 'fastcluster'

#> The following object is masked from 'package:stats':
#> 
#>     hclust

#>   site cluster_ward.D2
#> 1    A               1
#> 2    B               1
#> 3    C               1