Community detection on bipartite networks

Run a community detection algorithm on a bipartite network.
Algorithms available are greedy, betweenness, walktrap, louvain, spinglass, leading_eigen, label_prop (all from from igraph package), beckett and quanbimo from bipartite package, netcarto from rnetcarto R package and oslom and infomap that come from compiled codes (respectively from oslom.org and mapequation.org/infomap)
Algorithm can be weighted in case of a weighted bipartite network.

community(dat, algo = "greedy", weight = FALSE, input = "matrix", site = NULL, sp = NULL, ab = NULL)

Arguments

dat
a contingency matrix with sites as rows and species as columns.If 'algo="oslom"', a data.frame with three columns containing id1, id2 and a similarity metric (output of 'bioRgeo::simil()' function).

algo
An algorithm used to perfom the bipartite community detection. Available options are greedy, betweenness, walktrap, louvain, infomap, spinglass, leading_eigen, label_prop, netcarto, oslom, beckett and quanbimo.

weight
A boolean indicating whether the detection of communities has to be weighted or not.

input
A character string indicating whether the input format is a contingency matrix or a long-format data.frame.

site
a character string corresponding to the column with sites.

sp
a character string corresponding to the column with species.

ab
a character string corresponding to the column with abundances of species at sites.

n_runs
Numeric parameter setting the number of runs for the first hierarchical level. The default value is 10. The bigger, the slower the function but the more accurate the results.

t_param
Numeric parameter setting the probability value to consider a module as significant. Must be comprised between 0 and 1.

cp_param
Numeric parameter setting the resolution. Must be comprised between 0 and 1.

hr
Numeric parameter setting the number of hierarchical levels. Must be a positive integer.

oslom_id1
With oslom algorithm only. A character string corresponding to the column name of the first pair of sites.

oslom_id2
With oslom algorithm only. A character string corresponding to the column name of the second pair of sites..

oslom_proj
With oslom algorithm only. A character string corresponding to the column name of the similarity metric between each pair of sites.

Value

A data.frame with 2 columns: the sites and their associated bioregions.
Some algorithms (the ones from igraph package and infomap return a bioregion for both types of nodes from the bipartite network. This means that both sites and species are classified into bioregions.)

Details

Note

Examples

# Example dataset from bioRgeo:
data("virtual_sp")

#> Warning: data set 'virtual_sp' not found

sp_df <- virtual[[1]]

# fastgreedy algorithm from igraph
bip_df <- community(dat = sp_df[which(sp_df$pa > 0), ],
                         algo = "greedy",
                         weight = FALSE, input = "data.frame", site = "site",
                         sp = "sp", ab = NULL)

#> Error in sp_df$pa: $ operator is invalid for atomic vectors