Sites and species contributions

See Guimera et Amaral, Nature, 2005 for Cz coefficients. See Lenormand et al., Ecology and Evolution, 2019 for rho scores.

contribute(dat, sp_col, site_col, bioregion_col, bioregion_sp_col = NULL, ab = NULL)

Arguments

dat
a data.frame with each row indicating the presence of a species in a given site.

sp_col
a character string corresponding to the column with species.

site_col
a character string corresponding to the column with sites.

bioregion_col
a character string corresponding to the column with bioregions.

bioregion_sp_col
a character string corresponding to the column with bioregions of species. Set to NULL with algorithms that only assign a bioregion to sites.

ab
a logical indicating whether abundances should be taken into account.

Details

See Guimera et Amaral, Nature, 2005 for Cz coefficients. See Lenormand et al., Ecology and Evolution, 2019 for rho scores.

Note

Value

The input data.frame binded with 7 new columns: `bio_site` containing the bioregion of the given site, `bio_sp` the bioregion of the given species, `C_site` and `C_sp` having respectively the participation coefficient C score of the site and the species at each row, `z_site` and `z_sp` the same of the intra-module degree z coefficient and rho the species contribution.

Examples

ex <- data.frame(sites = c(rep("A", 2), rep("B", 3), rep("C", 2)),
species = c("a", "b", "a", "c", "d", "b", "d"),
count = c(10, 100, 1, 20, 50, 10, 20),
bioregion = c(rep("region1", 5), rep("region2", 2)))

scores <- contribute(dat = ex, sp_col = "species", site_col = "sites",
                 bioregion_col = "bioregion")

#> Warning: No bioregion provided for species. Each species will be assigned
#>             to the bioregion where it occurs the most.

#> Error in dat %>% group_by(sp, bio_site) %>% summarise(c = n()) %>% filter(row_number(desc(c)) ==     1) %>% rename(bio_sp = bio_site) %>% select(sp, bio_sp): impossible de trouver la fonction "%>%"

scores

#> Error in eval(expr, envir, enclos): objet 'scores' introuvable