Convert distance indices to similarity indices

This function converts a data.frame of distance indices (beta diversity) between sites to similarity indices.

distanceToSimilarity(distancedata)

Arguments

distancedata
the output object from similarityToDistance or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioRgeo.similarity

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - distance\)

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = "all")

#> Loading required package: Rcpp

simil

#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 a b c   Jaccard Jaccardturn  Sorensen   Simpson   A    B    C
#> 6  Site1 Site2 7 2 0 0.7777778   1.0000000 0.8750000 1.0000000 106  665 1071
#> 11 Site1 Site3 9 0 1 0.9000000   1.0000000 0.9473684 1.0000000 752   19 1319
#> 16 Site1 Site4 7 2 1 0.7000000   0.7777778 0.8235294 0.8750000  47  724   94
#> 21 Site1 Site5 7 2 1 0.7000000   0.7777778 0.8235294 0.8750000 132  639 1337
#> 12 Site2 Site3 7 0 3 0.7000000   1.0000000 0.8235294 1.0000000 291  886 1780
#> 17 Site2 Site4 6 1 2 0.6666667   0.7500000 0.8000000 0.8571429 106 1071   35
#> 22 Site2 Site5 5 2 3 0.5000000   0.5555556 0.6666667 0.7142857 379  798 1090
#> 18 Site3 Site4 8 2 0 0.8000000   1.0000000 0.8888889 1.0000000  49 2022   92
#> 23 Site3 Site5 8 2 0 0.8000000   1.0000000 0.8888889 1.0000000 151 1920 1318
#> 24 Site4 Site5 6 2 2 0.6000000   0.6000000 0.7500000 0.7500000  17  124 1452
#>          Bray  Brayturn    Euclidean
#> 6  0.10882957 0.1374838 0.0011226063
#> 11 0.52920479 0.9753567 0.0012971064
#> 16 0.10307018 0.3333333 0.0018551021
#> 21 0.11785714 0.1712062 0.0009549093
#> 12 0.17918719 0.2472387 0.0008596360
#> 17 0.16084977 0.7517730 0.0013549102
#> 22 0.28647014 0.3220051 0.0009767579
#> 18 0.04430380 0.3475177 0.0009435908
#> 23 0.08531073 0.1027910 0.0007292448
#> 24 0.02111801 0.1205674 0.0010800075

distances <- similarityToDistance(simil)
distances

#> Data.frame of distances between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of distance metrics:  7 
#> 
#> 
#>    Site1 Site2 a b c   Jaccard Jaccardturn   Sorensen   Simpson   A    B    C
#> 6  Site1 Site2 7 2 0 0.2222222   0.0000000 0.12500000 0.0000000 106  665 1071
#> 11 Site1 Site3 9 0 1 0.1000000   0.0000000 0.05263158 0.0000000 752   19 1319
#> 16 Site1 Site4 7 2 1 0.3000000   0.2222222 0.17647059 0.1250000  47  724   94
#> 21 Site1 Site5 7 2 1 0.3000000   0.2222222 0.17647059 0.1250000 132  639 1337
#> 12 Site2 Site3 7 0 3 0.3000000   0.0000000 0.17647059 0.0000000 291  886 1780
#> 17 Site2 Site4 6 1 2 0.3333333   0.2500000 0.20000000 0.1428571 106 1071   35
#> 22 Site2 Site5 5 2 3 0.5000000   0.4444444 0.33333333 0.2857143 379  798 1090
#> 18 Site3 Site4 8 2 0 0.2000000   0.0000000 0.11111111 0.0000000  49 2022   92
#> 23 Site3 Site5 8 2 0 0.2000000   0.0000000 0.11111111 0.0000000 151 1920 1318
#> 24 Site4 Site5 6 2 2 0.4000000   0.4000000 0.25000000 0.2500000  17  124 1452
#>         Bray   Brayturn Euclidean
#> 6  0.8911704 0.86251621  889.7842
#> 11 0.4707952 0.02464332  769.9468
#> 16 0.8969298 0.66666667  538.0539
#> 21 0.8821429 0.82879377 1046.2199
#> 12 0.8208128 0.75276126 1162.2831
#> 17 0.8391502 0.24822695  737.0563
#> 22 0.7135299 0.67799490 1022.7952
#> 18 0.9556962 0.65248227 1058.7814
#> 23 0.9146893 0.89720899 1370.2817
#> 24 0.9788820 0.87943262  924.9195

simil <- distanceToSimilarity(distances)
simil

#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 a b c   Jaccard Jaccardturn  Sorensen   Simpson   A    B    C
#> 6  Site1 Site2 7 2 0 0.7777778   1.0000000 0.8750000 1.0000000 106  665 1071
#> 11 Site1 Site3 9 0 1 0.9000000   1.0000000 0.9473684 1.0000000 752   19 1319
#> 16 Site1 Site4 7 2 1 0.7000000   0.7777778 0.8235294 0.8750000  47  724   94
#> 21 Site1 Site5 7 2 1 0.7000000   0.7777778 0.8235294 0.8750000 132  639 1337
#> 12 Site2 Site3 7 0 3 0.7000000   1.0000000 0.8235294 1.0000000 291  886 1780
#> 17 Site2 Site4 6 1 2 0.6666667   0.7500000 0.8000000 0.8571429 106 1071   35
#> 22 Site2 Site5 5 2 3 0.5000000   0.5555556 0.6666667 0.7142857 379  798 1090
#> 18 Site3 Site4 8 2 0 0.8000000   1.0000000 0.8888889 1.0000000  49 2022   92
#> 23 Site3 Site5 8 2 0 0.8000000   1.0000000 0.8888889 1.0000000 151 1920 1318
#> 24 Site4 Site5 6 2 2 0.6000000   0.6000000 0.7500000 0.7500000  17  124 1452
#>          Bray  Brayturn    Euclidean
#> 6  0.10882957 0.1374838 0.0011226063
#> 11 0.52920479 0.9753567 0.0012971064
#> 16 0.10307018 0.3333333 0.0018551021
#> 21 0.11785714 0.1712062 0.0009549093
#> 12 0.17918719 0.2472387 0.0008596360
#> 17 0.16084977 0.7517730 0.0013549102
#> 22 0.28647014 0.3220051 0.0009767579
#> 18 0.04430380 0.3475177 0.0009435908
#> 23 0.08531073 0.1027910 0.0007292448
#> 24 0.02111801 0.1205674 0.0010800075