Infomap community finding

This function finds communities in a (un)weighted (un)directed network based on the Infomap algorithm (https://github.com/mapequation/infomap, version 1.6.0).

infomap(
  net,
  weight = TRUE,
  bipartite = FALSE,
  nbmod = 0,
  markovtime = 1,
  seed = 1,
  twolevel = FALSE,
  directed = FALSE
)

Arguments

net
a two- or three-column data.frame representing a network with the two first columns as (un)directed links between pair of nodes and an optional third column indicating the weight of the link

weight
a boolean indicating if the weights should be considered if there is a third column

bipartite
a boolean indicating if the network is bipartite (see Details)

nbmod
penalize solutions the more they differ from this number (0 by default for no preferred number of modules)

markovtime
scales link flow to change the cost of moving between modules, higher values results in fewer modules (default is 1)

seed
for the random number generator

twolevel
a boolean indicating if the algorithm should optimize a two-level partition of the network (default is multi-level)

directed
a boolean indicating if the network is directed (from column 1 to column 2)

Value

A data.frame providing one partition by hierarchical level.

Details

Infomap is a network clustering algorithm based on the Map equation proposed in (Rosvall and Bergstrom 2008) that finds communities in (un)weighted and (un)directed networks. Infomap has two ways to deal with bipartite networks. The first possibility is to consider the bipartite network as unipartite network. The second possibility is to set the bipartite argument to TRUE in order to approximate a two-step random walker (see https://www.mapequation.org/infomap/ for more information). This function is based on the 1.6.0 C++ version of Infomap (https://github.com/mapequation/infomap/releases).

References

Rosvall M, Bergstrom CT (2008). “Maps of random walks on complex networks reveal community structure.” Proceedings of the National Academy of Sciences, 105(4), 1118--1123.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=spproject(comat,metric="Simpson")

#> Loading required package: Rcpp

com=infomap(net)

#> Error in file(file, ifelse(append, "a", "w")): invalid 'description' argument