Louvain community finding

This function finds communities in a (un)weighted undirected network based on the Louvain algorithm.

louvain(net, weight = TRUE, lang = "Cpp", q = 0, c = 0.5, k = 1)

Arguments

net
a two- or three-column data.frame representing a network with the two first columns as (un)directed links between pair of nodes and an optional third column indicating the weight of the link

weight
a boolean indicating if the weights should be considered if there is a third column

lang
a string indicating what version of Louvain should be used (igraph or Cpp, see Details)

q
the quality function used to compute partition of the graph (modularity is chosen by default, see Details)

c
the parameter for the Owsinski-Zadrozny quality function (between 0 and 1, 0.5 is chosen by default)

k
the kappa_min value for the Shi-Malik quality function (it must be > 0, 1 is chosen by default)

Value

A data.frame providing one community by node.

Details

Louvain is a network community detection algorithm proposed in (Blondel et al. 2008) . This function proposed two implementations of the function (parameter lang): the igraph implementation (cluster_louvain()) and the C++ implementation (https://sourceforge.net/projects/louvain/, version 0.3). The latest offers the possibility to choose among several quality functions, q = 0 for the classical Newman-Girvan criterion (also called "Modularity"), 1 for the Zahn-Condorcet criterion, 2 for the Owsinski-Zadrozny criterion (you should specify the value of the parameter with option -c), 3 for the Goldberg Density criterion, 4 for the A-weighted Condorcet criterion, 5 for the Deviation to Indetermination criterion, 6 for the Deviation to Uniformity criterion, 7 for the Profile Difference criterion, 8 for the Shi-Malik criterion (you should specify the value of kappa_min with option -k) and 9 for the Balanced Modularity criterion.

References

Blondel VD, Guillaume JL, Lambiotte R, Mech ELJS (2008). “Fast unfolding of communities in large networks.” J. Stat. Mech, P10008.

Author

Pierre Denelle pierre.denelle@gmail.com Maxime Lenormand maxime.lenormand@inrae.fr Boris Leroy leroy.boris@gmail.com

Examples

comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=spproject(comat,metric="Simpson")
com=louvain(net)

#> Error in file(file, ifelse(append, "a", "w")): invalid 'description' argument