OSLOM community finding

This function finds communities in a (un)weighted (un)directed network based on the OSLOM algorithm (http://oslom.org/, version 2.4).

oslom(
  net,
  weight = TRUE,
  reassign = "no",
  r = 10,
  hr = 50,
  seed = 1,
  t = 0.1,
  cp = 0.5,
  directed = FALSE
)

Arguments

net
a two- or three-column data.frame representing a network with the two first columns as (un)directed links between pair of nodes and an optional third column indicating the weight of the link

weight
a boolean indicating if the weights should be considered if there is a third column

reassign
a string indicating if the nodes belonging to several community should be reassign and what method should be used (see Details)

r
the number of runs for the first hierarchical level (10 by default)

hr
the number of runs for the higher hierarchical level (50 by default, 0 if you are not interested in hierarchies)

seed
for the random number generator

t
the p−value, the default value is 0.10, increase this value you to get more modules

cp
kind of resolution parameter used to decide between taking some modules or their union (default value is 0.5, bigger value leads to bigger clusters)

directed
a boolean indicating if the network is directed (from column 1 to column 2)

Value

A data.frame providing one or several modules (according to the chosen option(s)) for each node. If reassign = simil or reassign = random only one column by hierarchical level will be provided. If reassign = no the number of columns will depend on the number of overlapping modules (up to three). A value of 0 indicates that no module were found (i.e. non-overlapping nodes).

Details

OSLOM is a network community detection algorithm proposed in (Lancichinetti et al. 2011) that finds statistically significant (overlapping) communities in (un)weighted and (un)directed networks. Since a node may belong to several communities we propose two methods to reassign the 'ovelapping' nodes randomly reassign = 'random' or based on the closest candidate community reassign = 'simil' (only for weighted networks, in this case the closest candidite community is determined with the average similarity). By default reassign = 'no' and all the information is provided. This function is based on the 2.4 C++ version of OSLOM (http://www.oslom.org/software.htm).

References

Lancichinetti A, Radicchi F, Ramasco JJ, Fortunato S (2011). “Finding Statistically Significant Communities in Networks.” PLOS ONE, 6(4), e18961.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=spproject(comat,metric="Simpson")
com=oslom(net)

#> Error in file(file, ifelse(append, "a", "w")): invalid 'description' argument