Similarity matrix

Project a contingency matrix into a similarity data.frame, the similarity being based upon the sites each pair of species has in common.

simil(dat, metric = "simpson", input = "matrix", output = "data.frame", site = NULL, sp = NULL, ab = NULL, weight = FALSE)

Arguments

dat
a contingency matrix with sites as rows and species as columns.

metric
a character string indicating which similarity metric to chose among simpson, jaccard, sorensen, whittaker, euclidean and bray.

input
a character string equals to 'matrix' or 'data.frame' indicating whether the input of the function is long-format data.frame or a wide format matrix.

output
a character string indicating whether the output of the function should be a symmetrical matrix, a long-format data.frame or a distance object.

site
If input = 'data.frame', the column name of the sites.

sp
If input = 'data.frame', the column name of the species.

ab
If input = 'data.frame' and species have abundances, the column name of the abundances of species per site.

weight
a boolean indicating whether abundances should be taken into account.

Details

With a the number of species shared by a pair of sites, b species only present in the first site and c species only present in the second site.

\(Simpson = 1 - min(b, c)/(a + min(b, c))\)

\(Sorensen = 1 - (b+c)/(2a + b + c)\)

\(Jaccard = 1 - a/(a + b + c)\)

\(Whittaker = 1 - (a + b + c)/((2a + b + c)/2)\)

Euclidean computes the Euclidean distance between each pair of site.

If abundances data are available, Bray-Curtis dissimilarity can also be computed. It computes following this equation: \(Bray-Curtis = 1 - 2C_{ij}/(S_i + S_j)\) with \(2C_{ij}\) the sum of the lesser values for common species shared by a pair of sites. \(S_{i}\) and \(S_{j}\) are the total number of specimens counted at both sites \(i\) and \(j\).

Value

If output = 'data.frame', a data.frame with 3 columns. The first two columns, id1 and id2, correspond to each pair of sites. The third column indicates the similarity value for the chosen metric between each pair of site.

If output = 'matrix', a matrix with n_sites rows and n_sites columns. Values of the matrix indicate the similarity value for the chosen metric between each pair of site.

If output = 'dist', a dist object based on the symmetrical matrix.

Note

Examples

ex <- data.frame(sites = c(rep("A", 2), rep("B", 3), rep("C", 2)),
species = c("a", "b", "a", "c", "d", "b", "d"),
count = c(10, 100, 1, 20, 50, 10, 20))

ex_mat <- contingency(dat = ex, sp = "species", site = "sites", ab = "count",
weight = FALSE)
simil(ex_mat, metric = "simpson", input = "matrix")

#>   id1 id2 simpson
#> 1   A   B     0.5
#> 2   A   C     0.5
#> 3   B   C     0.5