Convert similarity indices to distance indices

This function converts a data.frame of similarity indices between sites to distance indices (= beta diversity).

similarityToDistance(similaritydata)

Arguments

similaritydata
the output object from spproject or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioRgeo.distance

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into distances with the following formula:

\(distance = 1 - similarity\)

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = "all")
simil

#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2  a b c   Jaccard Jaccardturn  Sorensen   Simpson   A   B    C
#> 6  Site1 Site2  9 0 1 0.9000000   1.0000000 0.9473684 1.0000000 431 693  266
#> 11 Site1 Site3  8 1 0 0.8888889   1.0000000 0.9411765 1.0000000 509 615  432
#> 16 Site1 Site4  8 1 1 0.8000000   0.8000000 0.8888889 0.8888889 321 803 1144
#> 21 Site1 Site5  9 0 1 0.9000000   1.0000000 0.9473684 1.0000000 221 903 2597
#> 12 Site2 Site3  8 2 0 0.8000000   1.0000000 0.8888889 1.0000000 183 514  758
#> 17 Site2 Site4  9 1 0 0.9000000   1.0000000 0.9473684 1.0000000 217 480 1248
#> 22 Site2 Site5 10 0 0 1.0000000   1.0000000 1.0000000 1.0000000 445 252 2373
#> 18 Site3 Site4  7 1 2 0.7000000   0.7777778 0.8235294 0.8750000 304 637 1161
#> 23 Site3 Site5  8 0 2 0.8000000   1.0000000 0.8888889 1.0000000 341 600 2477
#> 24 Site4 Site5  9 0 1 0.9000000   1.0000000 0.9473684 1.0000000 508 957 2310
#>         Bray  Brayturn    Euclidean
#> 6  0.4733663 0.6183644 0.0020496461
#> 11 0.4929782 0.5409139 0.0021979139
#> 16 0.2479722 0.2855872 0.0010642398
#> 21 0.1121258 0.1966192 0.0006638989
#> 12 0.2234432 0.2625538 0.0019429590
#> 17 0.2007401 0.3113343 0.0011429289
#> 22 0.2532006 0.6384505 0.0007750144
#> 18 0.2527016 0.3230606 0.0011662746
#> 23 0.1814312 0.3623804 0.0007614332
#> 24 0.2372169 0.3467577 0.0006580866

distances <- similarityToDistance(simil)
distances

#> Data.frame of distances between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of distance metrics:  7 
#> 
#> 
#>    Site1 Site2  a b c   Jaccard Jaccardturn   Sorensen   Simpson   A   B    C
#> 6  Site1 Site2  9 0 1 0.1000000   0.0000000 0.05263158 0.0000000 431 693  266
#> 11 Site1 Site3  8 1 0 0.1111111   0.0000000 0.05882353 0.0000000 509 615  432
#> 16 Site1 Site4  8 1 1 0.2000000   0.2000000 0.11111111 0.1111111 321 803 1144
#> 21 Site1 Site5  9 0 1 0.1000000   0.0000000 0.05263158 0.0000000 221 903 2597
#> 12 Site2 Site3  8 2 0 0.2000000   0.0000000 0.11111111 0.0000000 183 514  758
#> 17 Site2 Site4  9 1 0 0.1000000   0.0000000 0.05263158 0.0000000 217 480 1248
#> 22 Site2 Site5 10 0 0 0.0000000   0.0000000 0.00000000 0.0000000 445 252 2373
#> 18 Site3 Site4  7 1 2 0.3000000   0.2222222 0.17647059 0.1250000 304 637 1161
#> 23 Site3 Site5  8 0 2 0.2000000   0.0000000 0.11111111 0.0000000 341 600 2477
#> 24 Site4 Site5  9 0 1 0.1000000   0.0000000 0.05263158 0.0000000 508 957 2310
#>         Bray  Brayturn Euclidean
#> 6  0.5266337 0.3816356  486.8891
#> 11 0.5070218 0.4590861  453.9769
#> 16 0.7520278 0.7144128  938.6378
#> 21 0.8878742 0.8033808 1505.2535
#> 12 0.7765568 0.7374462  513.6789
#> 17 0.7992599 0.6886657  873.9451
#> 22 0.7467994 0.3615495 1289.2986
#> 18 0.7472984 0.6769394  856.4310
#> 23 0.8185688 0.6376196 1312.3128
#> 24 0.7627831 0.6532423 1518.5569