Compute similarity metrics between sites based on species composition

This function creates a data.frame where each row provides one or several similarity metric(s) between each pair of sites from a co-occurence matrix with sites as rows and species as columns.

spproject(comat, metric = "Simpson", method = "prodmat")

Arguments

comat
a co-occurence matrix with sites as rows and species as columns

metric
a vector of character(s) indicating which similarity metric(s) to chose (see Details). If "all" is specified, then all metrics will be calculated.

method
a string indicating what method should be used to compute abc (see Details), method = "prodmat" by default is more efficient but can greedy in memory and method="loops" is less efficient but less greedy in memory

Value

A data.frame providing one or several similarity metric(s) between each pair of sites. The two first columns represents each pair of sites. One column per similarity metric provided in metric except for the metric abc and ABC that are stored in three columns (one for each letter).

Details

With a the number of species shared by a pair of sites, b species only present in the first site and c species only present in the second site.

\(Jaccard = 1 - (b + c) / (a + b + c)\)

\(Jaccardturn = 1 - 2min(b, c) / (a + 2min(b, c))\) (Baselga 2012)

\(Sorensen = 1 - (b + c) / (2a + b + c)\)

\(Simpson = 1 - min(b, c) / (a + min(b, c))\)

If abundances data are available, Bray-Curtis and its turnover component can also be computed with the following equation:

\(Bray = 1 - (B + C) / (2A + B + C)\)

\(Brayturn = 1 - min(B, C)/(A + min(B, C))\) (Baselga 2013)

with A the sum of the lesser values for common species shared by a pair of sites. B and C are the total number of specimens counted at both sites minus A.

Euclidean computes the Euclidean similarity between each pair of site following this equation:

\(Euclidean = 1 / (1 + dij)\)

Where dij is the Euclidean distance between site i and site j in terms of species composition.

References

Baselga A (2012). “The Relationship between Species Replacement, Dissimilarity Derived from Nestedness, and Nestedness.” Global Ecology and Biogeography, 21(12), 1223--1232. Baselga A (2013). “Separating the two components of abundance-based dissimilarity: balanced changes in abundance vs. abundance gradients.” Methods in Ecology and Evolution, 4(6), 552--557.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = c("abc", "ABC", "Simpson", "Brayturn"))
simil

#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  2 
#> 
#> 
#>    Site1 Site2 a b c   Simpson   A   B    C  Brayturn
#> 6  Site1 Site2 8 1 1 0.8888889  88 909  593 0.1292217
#> 11 Site1 Site3 8 1 1 0.8888889 161 836 1009 0.1614845
#> 16 Site1 Site4 7 2 1 0.8750000 374 623  839 0.3751254
#> 21 Site1 Site5 8 1 1 0.8888889 379 618  799 0.3801404
#> 12 Site2 Site3 9 0 0 1.0000000 425 256  745 0.6240822
#> 17 Site2 Site4 8 1 0 1.0000000 268 413  945 0.3935389
#> 22 Site2 Site5 8 1 1 0.8888889  69 612 1109 0.1013216
#> 18 Site3 Site4 8 1 0 1.0000000 469 701  744 0.4008547
#> 23 Site3 Site5 8 1 1 0.8888889 298 872  880 0.2547009
#> 24 Site4 Site5 7 1 2 0.8750000 480 733  698 0.4074703

simil <- spproject(comat, metric = "all")
simil

#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 a b c   Jaccard Jaccardturn  Sorensen   Simpson   A   B    C
#> 6  Site1 Site2 8 1 1 0.8000000   0.8000000 0.8888889 0.8888889  88 909  593
#> 11 Site1 Site3 8 1 1 0.8000000   0.8000000 0.8888889 0.8888889 161 836 1009
#> 16 Site1 Site4 7 2 1 0.7000000   0.7777778 0.8235294 0.8750000 374 623  839
#> 21 Site1 Site5 8 1 1 0.8000000   0.8000000 0.8888889 0.8888889 379 618  799
#> 12 Site2 Site3 9 0 0 1.0000000   1.0000000 1.0000000 1.0000000 425 256  745
#> 17 Site2 Site4 8 1 0 0.8888889   1.0000000 0.9411765 1.0000000 268 413  945
#> 22 Site2 Site5 8 1 1 0.8000000   0.8000000 0.8888889 0.8888889  69 612 1109
#> 18 Site3 Site4 8 1 0 0.8888889   1.0000000 0.9411765 1.0000000 469 701  744
#> 23 Site3 Site5 8 1 1 0.8000000   0.8000000 0.8888889 0.8888889 298 872  880
#> 24 Site4 Site5 7 1 2 0.7000000   0.7777778 0.8235294 0.8750000 480 733  698
#>          Bray  Brayturn    Euclidean
#> 6  0.10488677 0.1292217 0.0013444229
#> 11 0.14859252 0.1614845 0.0010466466
#> 16 0.33846154 0.3751254 0.0014106396
#> 21 0.34850575 0.3801404 0.0012673711
#> 12 0.45921124 0.6240822 0.0021812812
#> 17 0.28299894 0.3935389 0.0014563119
#> 22 0.07423346 0.1013216 0.0011428400
#> 18 0.39362149 0.4008547 0.0012581616
#> 23 0.25383305 0.2547009 0.0009972849
#> 24 0.40150565 0.4074703 0.0013021894