Compute similarity metrics between sites based on species composition

This function creates a data.frame where each row provides one or several similarity metric(s) between each pair of sites from a co-occurence matrix with sites as rows and species as columns.

spproject(comat, metric = "Simpson", method = "prodmat")

Arguments

comat
a co-occurence matrix with sites as rows and species as columns

metric
a vector of character(s) indicating which similarity metric(s) to chose (see Details)

method
a string indicating what method should be used to compute abc (see Details), method = "prodmat" by default is more efficient but can greedy in memory and method="loops" is less efficient but less greedy in memory

Value

A data.frame providing one or several similarity metric(s) between each pair of sites. The two first columns represents each pair of sites. One column per similarity metric provided in metric except for the metric abc and ABC that are stored in three columns (one for each letter).

Details

With a the number of species shared by a pair of sites, b species only present in the first site and c species only present in the second site.

\(Jaccard = 1 - (b+c)/(a + b + c)\)

\(Jaccardturn = 1 - 2min(b, c)/(a + 2min(b, c))\) (Baselga 2012)

\(Sorensen = 1 - (b+c)/(2a + b + c)\)

\(Simpson = 1 - min(b, c)/(a + min(b, c))\)

If abundances data are available, Bray-Curtis and its turnover component can also be computed with the following equation:

\(Bray = 1- (B+C)/(2A + B + C)\)

\(Brayturn = 1 - min(B, C)/(A + min(B, C))\) (Baselga 2013)

with A the sum of the lesser values for common species shared by a pair of sites. B and C are the total number of specimens counted at both sites minus A.

Euclidean computes the Euclidean similarity between each pair of site following this equation:

\(Euclidean = 1 /(1 + dij)\)

Where dij is the Euclidean distance between site i and site j in terms of species composition.

References

Baselga A (2012). “The Relationship between Species Replacement, Dissimilarity Derived from Nestedness, and Nestedness.” Global Ecology and Biogeography, 21(12), 1223--1232. Baselga A (2013). “Separating the two components of abundance-based dissimilarity: balanced changes in abundance vs. abundance gradients.” Methods in Ecology and Evolution, 4(6), 552--557.

Author

Pierre Denelle (pierre.denelle@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Boris Leroy (leroy.boris@gmail.com)

Examples

comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

simil=spproject(comat,metric=c("abc","ABC","Simpson","Brayturn"))